+++ screen2.0/src/app/applets/argo/genes/geneTable.tsx
