+++ screen2.0/src/app/applets/argo/sequence/sequenceTable.tsx
